Refine ADR-012: factor RESERVED_NAMES, extract _regression, align docs
- Soften ADR-012 + CLAUDE.md to describe what's actually enforced:
  PLC2701 is the hard rule; underscores on function names are an
  optional convention.
- Split _utils.RESERVED_NAMES into RESERVED_COORD_NAMES (coord-name
  collisions) and RESERVED_COMPARER_VAR_NAMES (+Observation slot),
  expressing the superset relationship in code instead of by
  shadowed in-function constants.
- Extract linear_regression from _utils.py into its own
  single-purpose _regression.py, leaving _utils.py as the
  name-handling helpers module.
